🐛 fix output from DurationScroll on fast scroll
