+++ src/index.ts
@@ -1,6 +1,12 @@
 export {
+    default as DurationPickerModal,
+    DurationPickerModalProps,
+} from "./components";
+
+export {
     default as DurationPicker,
     DurationPickerProps,
 } from "./components/DurationPicker";
 
+export { CustomDurationPickerModalStyles } from "./components/DurationPickerModal.styles";
 export { CustomDurationPickerStyles } from "./components/DurationPicker/DurationPicker.styles";
